+++ lib/jazzy/source_declaration/access_control_level.rb
@@ -5,26 +5,51 @@ module Jazzy
 
       attr_reader :level
 
-      def initialize(declaration_string)
-        if declaration_string =~ /private\ /
+      def initialize(accessibility)
+        if accessibility == 'source.lang.swift.accessibility.private'
           @level = :private
-        elsif declaration_string =~ /public\ /
-          @level = :public
-        else
+        elsif accessibility == 'source.lang.swift.accessibility.internal'
           @level = :internal
+        elsif accessibility == 'source.lang.swift.accessibility.public'
+          @level = :public
+        end
+      end
+
+      def self.from_doc(doc)
+        accessibility = doc['key.accessibility']
+        if accessibility
+          acl = new(accessibility)
+          if acl
+            return acl
+          end
+        end
+        acl = from_explicit_declaration(doc['key.parsed_declaration'])
+        if acl
+          return acl
+        end
+        AccessControlLevel.public # We don't know what ACL this declaration is
+      end
+
+      def self.from_explicit_declaration(declaration_string)
+        if declaration_string =~ /private\ /
+          return AccessControlLevel.private
+        elsif declaration_string =~ /public\ /
+          return AccessControlLevel.public
+        elsif declaration_string =~ /internal\ /
+          return AccessControlLevel.internal
         end
       end
 
       def self.private
-        AccessControlLevel.new('private ')
+        AccessControlLevel.new('source.lang.swift.accessibility.private')
       end
 
       def self.internal
-        AccessControlLevel.new('internal ')
+        AccessControlLevel.new('source.lang.swift.accessibility.internal')
       end
 
       def self.public
-        AccessControlLevel.new('public ')
+        AccessControlLevel.new('source.lang.swift.accessibility.public')
       end
 
       LEVELS = {